+++ tests/data/test_breadth_first_walker.py
@@ -18,7 +18,7 @@ import pytest
 import networkx as nx
 from stellargraph.data.explorer import SampledBreadthFirstWalk
 from stellargraph.core.graph import StellarDiGraph
-from ..test_utils.graph_fixtures import create_test_graph
+from ..test_utils.graphs import create_test_graph
 
 
 def expected_bfw_size(n_size):
