+++ tests/layer/test_cluster_gcn.py
@@ -29,7 +29,7 @@ import pandas as pd
 import numpy as np
 from tensorflow import keras
 import pytest
-from ..test_utils.graph_fixtures import create_graph_features, create_stellargraph
+from ..test_utils.graphs import create_graph_features, create_stellargraph
 
 
 def test_ClusterGraphConvolution_config():
